Improve Qemu Workaround
Qemu appears to have a memory conflict with the Title Banner PixelData.

Until a proper solution is found, a memory leak has been permitted when 
ConsoleControl is absent (such as when Qemu is active).

This commit limits the permitted leak to the Title Banner as opposed to 
all images as was previously the case (3ff6796)